Back End/requirements.txt
Flask
gunicorn
flask-cors
google-generativeai
